+++ monopyly/CHANGELOG.md
@@ -229,6 +229,7 @@
 - Remove Python version requirement (allow Python versions after 3.10)
 - Update dependencies (including using ruff in place of _Black_ and _isort_)
 - Refactor to include ruff-based linting checks
+- Set 'Credit payments' to be a default tag for all users
 
 
 <a name="bottom" id="bottom"></a>